==> src/14_Class_Modeling.ipynb <==
# !pip install skops

# from tpot import TPOTClassifier
# from imblearn.ensemble import BalancedBaggingClassifier
# from sklearn.ensemble import ExtraTreesClassifier
# from sklearn.model_selection import train_test_split
# from sklearn.metrics import classification_report

import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

from sklearn.metrics import confusion_matrix
from sklearn.metrics import classification_report
from sklearn.model_selection import train_test_split
from sklearn.decomposition import PCA
from sklearn.decomposition import KernelPCA
from sklearn.ensemble import RandomForestClassifier, ExtraTreesClassifier, GradientBoostingClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import MinMaxScaler
from sklearn.model_selection import learning_curve
from sklearn.model_selection import ShuffleSplit
 
from pylab import rcParams
from imblearn.pipeline import Pipeline
import statsmodels.api as sm

from sklearn.datasets import load_iris
from sklearn.preprocessing import StandardScaler, MinMaxScaler
from sklearn.pipeline import make_pipeline
from sklearn.model_selection import cross_val_score, GridSearchCV

# import skops.io as sio

import sklearn

print(sklearn.__version__)

import pandas as pd
print(pd.__version__)

import sys
import pickle

print("Python version:", sys.version)
print("Pickle version (protocol):", pickle.HIGHEST_PROTOCOL)


df = pd.read_csv('../data/df_category.csv',sep=',', encoding='latin-1', index_col=0, low_memory=False)
df_edit = pd.read_csv('../data_edit/df_edit_category.csv',sep=',', encoding='latin-1', index_col=0, low_memory=False)
df_2017 = pd.read_csv('../data/df_2017_category.csv',sep=',', encoding='latin-1', index_col=0, low_memory=False)
df_edit_2017 = pd.read_csv('../data_edit/df_edit_2017_category.csv',sep=',', encoding='latin-1', index_col=0, low_memory=False)

# df_without = pd.read_csv('../data/df_category_without_balance.csv',sep=',', encoding='latin-1', index_col=0, low_memory=False)

# def histogram(df, name):
#     # Orden específico de las categorías
#     orden = ['quince', 'veinte', 'veinticinco', 'treinta', 'treinta y cinco', 'cuarenta', 'cuarenta y cinco', 'cincuenta']
    
#     plt.figure(figsize=(10,6))
#     # Crear el histograma
#     df['dosis_coagulante'].value_counts().reindex(orden).plot(kind='bar')

#     # Personalizar el gráfico
#     plt.title(f'Cantidad de datos para cada categoría de dosis de coagulante de la\n{name}')
#     plt.xlabel('Categorías')
#     plt.ylabel('Frecuencia')
#     plt.xticks(rotation=45)  # Rota las etiquetas del eje x si es necesario

#     # Mostrar el gráfico
#     plt.show()

# histogram(df, 'base de datos sin edición del 2013-2022.')

# histogram(df_edit, 'base de datos con edición del 2013-2022.')

# histogram(df_2017, 'base de datos sin edición del 2017-2022.')

# histogram(df_edit_2017, 'base de datos con edición del 2013-2022.')

# def run_other_algorithms(dataframes, target_col='dosis_coagulante'):
#     base_estimators = [RandomForestClassifier(), ExtraTreesClassifier(), KNeighborsClassifier(), 
#                        LogisticRegression(max_iter=1000),DecisionTreeClassifier(), GradientBoostingClassifier()]
#     base_estimators_names = ['Random Forest', 'Extra Trees', 'K-Nearest Neighbors', 'Logistic Regression',
#                              'Decision Tree', 'Gradient Boosting']

#     for i, df in enumerate(dataframes):
#         print(f"\nTrabajando con el dataframe número {i+1}")
        
#         # Preparar datos
#         y = df[target_col]
#         X = df.drop(target_col, axis=1)
#         X_train, X_test, y_train, y_test = train_test_split(X, y, train_size=0.7)

#         # Ejecutar modelos
#         for estimator, name in zip(base_estimators, base_estimators_names):
#             print(f"\nEjecutando algoritmo {name}")
#             a_est = estimator
#             a_est.fit(X_train, y_train)
#             pred_y = a_est.predict(X_test)
#             mostrar_resultados(y_test, pred_y)

# def mostrar_resultados(y_test, pred_y):
#     conf_matrix = confusion_matrix(y_test, pred_y)
#     LABELS = np.unique(y_test)
#     plt.figure(figsize=(12, 12))
#     sns.heatmap(conf_matrix, xticklabels=LABELS, yticklabels=LABELS, annot=True, fmt="d");
#     plt.title("Confusion matrix")
#     plt.ylabel('True class')
#     plt.xlabel('Predicted class')
#     plt.show()
#     print (classification_report(y_test, pred_y))

# dataframes = [df, df_edit, df_2017, df_edit_2017]
# run_other_algorithms(dataframes)

## Prueba de sobreajuste

# def run_test(dataframes, target_col='dosis_coagulante'):
#     base_estimators = [RandomForestClassifier(), ExtraTreesClassifier(), KNeighborsClassifier(), 
#                        LogisticRegression(max_iter=1000), DecisionTreeClassifier(), GradientBoostingClassifier()]
#     base_estimators_names = ['Random Forest', 'Extra Trees', 'K-Nearest Neighbors', 'Logistic Regression',
#                              'Decision Tree', 'Gradient Boosting']

#     for i, df in enumerate(dataframes):
#         print(f"\nTrabajando con el dataframe número {i+1}")
        
#         # Preparar datos
#         y = df[target_col]
#         X = df.drop(target_col, axis=1)
#         X_train, X_test, y_train, y_test = train_test_split(X, y, train_size=0.7)

#         # Ejecutar modelos
#         for estimator, name in zip(base_estimators, base_estimators_names):
#             print(f"\nEjecutando algoritmo {name}")
#             a_est = estimator
#             a_est.fit(X_train, y_train)

#             # Predicciones en el conjunto de entrenamiento
#             train_preds = a_est.predict(X_train)
#             # Predicciones en el conjunto de prueba
#             test_preds = a_est.predict(X_test)
            
#             print("Informe de clasificación para el conjunto de entrenamiento:")
#             mostrar_resultados(y_train, train_preds)

#             print("Informe de clasificación para el conjunto de prueba:")
#             mostrar_resultados(y_test, test_preds)

# def mostrar_resultados(y_true, y_pred):
#     conf_matrix = confusion_matrix(y_true, y_pred)
#     LABELS = np.unique(y_true)
#     plt.figure(figsize=(12, 12))
#     sns.heatmap(conf_matrix, xticklabels=LABELS, yticklabels=LABELS, annot=True, fmt="d");
#     plt.title("Confusion matrix")
#     plt.ylabel('True class')
#     plt.xlabel('Predicted class')
#     plt.show()
#     print (classification_report(y_true, y_pred))

# dataframes = [df, df_edit, df_2017, df_edit_2017]
# run_test(dataframes)

# dataframes = [df_without]
# run_test(dataframes)

Tras un cuidadoso análisis de los resultados, se ha determinado que los algoritmos de Random Forest y Extra Trees ofrecen el rendimiento más óptimo para los conjuntos de datos disponibles. En consecuencia, procederemos a realizar una evaluación más detallada de estos algoritmos, aplicando ajustes específicos a los hiperparámetros para cada conjunto de datos, con la finalidad de disminuir el sobreajuste de los modelos.

# def train_and_evaluate_rf(df, n_estimators, max_depth, target='dosis_coagulante'):
#     # Definir X e y
#     y = df[target]
#     X = df.drop(target, axis=1)

#     # Divide tus datos en conjuntos de entrenamiento y prueba
#     X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

#     # Crea un nuevo clasificador con los mejores hiperparámetros
#     clf = make_pipeline(RandomForestClassifier(n_estimators=n_estimators, max_depth=max_depth, random_state=42))

#     # Entrena el clasificador en el conjunto de entrenamiento
#     clf.fit(X_train, y_train)

#     # Prueba el clasificador en el conjunto de prueba
#     y_pred = clf.predict(X_test)

#     # Predicciones en el conjunto de entrenamiento
#     train_preds = clf.predict(X_train)

#     # Imprime los informes de clasificación para el entrenamiento y prueba
#     print("Informe de clasificación para el conjunto de entrenamiento:")
#     print(classification_report(y_train, train_preds))
    
#     print("Informe de clasificación para el conjunto de prueba:")
#     print(classification_report(y_test, y_pred))

# def train_and_evaluate_et(df, n_estimators, max_depth, min_samples_split, target='dosis_coagulante'):
#     # Definir X e y
#     y = df[target]
#     X = df.drop(target, axis=1)

#     # Divide tus datos en conjuntos de entrenamiento y prueba
#     X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

#     # Crea un nuevo clasificador con los mejores hiperparámetros
#     clf = make_pipeline(ExtraTreesClassifier(n_estimators=n_estimators, max_depth=max_depth, min_samples_split=min_samples_split, random_state=42))

#     # Entrena el clasificador en el conjunto de entrenamiento
#     clf.fit(X_train, y_train)

#     # Prueba el clasificador en el conjunto de prueba
#     y_pred = clf.predict(X_test)

#     # Predicciones en el conjunto de entrenamiento
#     train_preds = clf.predict(X_train)

#     # Imprime los informes de clasificación para el entrenamiento y prueba
#     print("Informe de clasificación para el conjunto de entrenamiento:")
#     print(classification_report(y_train, train_preds))
    
#     print("Informe de clasificación para el conjunto de prueba:")
#     print(classification_report(y_test, y_pred))

# def evaluate_model_cv_rf(df, n_estimators, max_depth , target_col='dosis_coagulante'):
#     y = df[target_col]
#     X = df.drop(target_col, axis=1)
    
#     # Crear un pipeline para estandarizar los datos y luego aplicar el clasificador
#     clf = make_pipeline(RandomForestClassifier(n_estimators=n_estimators, max_depth=max_depth, random_state=42))

#     # Realiza la validación cruzada
#     scores = cross_val_score(clf, X, y, cv=5)  
#     print("Scores de cada iteración de la validación cruzada: ", scores)
#     print("Media de los scores: ", scores.mean())
    
#     recall_scores = cross_val_score(clf, X, y, cv=5, scoring='recall_weighted')
#     print("Recall de cada iteración de la validación cruzada: ", recall_scores)
#     print("Media de los recalls: ", recall_scores.mean())

#     precision_scores = cross_val_score(clf, X, y, cv=5, scoring='precision_weighted')
#     print("Precision ponderada de cada iteración de la validación cruzada: ", precision_scores)
#     print("Media de las precisiones ponderadas: ", precision_scores.mean())

#     f1_scores = cross_val_score(clf, X, y, cv=5, scoring='f1_weighted')
#     print("F1-score de cada iteración de la validación cruzada: ", f1_scores)
#     print("Media de los F1-scores: ", f1_scores.mean())

#     accuracy_scores = cross_val_score(clf, X, y, cv=5, scoring='accuracy')
#     print("Accuracy de cada iteración de la validación cruzada: ", accuracy_scores)
#     print("Media de los accuracy: ", accuracy_scores.mean())

# def evaluate_model_cv_et(df, n_estimators, max_depth, min_samples_split, target_col='dosis_coagulante'):
#     y = df[target_col]
#     X = df.drop(target_col, axis=1)
    
#     # Crear un pipeline para estandarizar los datos y luego aplicar el clasificador
#     clf = make_pipeline(ExtraTreesClassifier(n_estimators=n_estimators, max_depth=max_depth, min_samples_split=min_samples_split, random_state=42))

#     # Realiza la validación cruzada
#     scores = cross_val_score(clf, X, y, cv=5)  

#     # Realiza la validación cruzada
#     scores = cross_val_score(clf, X, y, cv=5)  
#     print("Scores de cada iteración de la validación cruzada: ", scores)
#     print("Media de los scores: ", scores.mean())
    
#     recall_scores = cross_val_score(clf, X, y, cv=5, scoring='recall_weighted')
#     print("Recall de cada iteración de la validación cruzada: ", recall_scores)
#     print("Media de los recalls: ", recall_scores.mean())

#     precision_scores = cross_val_score(clf, X, y, cv=5, scoring='precision_weighted')
#     print("Precision ponderada de cada iteración de la validación cruzada: ", precision_scores)
#     print("Media de las precisiones ponderadas: ", precision_scores.mean())

#     f1_scores = cross_val_score(clf, X, y, cv=5, scoring='f1_weighted')
#     print("F1-score de cada iteración de la validación cruzada: ", f1_scores)
#     print("Media de los F1-scores: ", f1_scores.mean())

#     accuracy_scores = cross_val_score(clf, X, y, cv=5, scoring='accuracy')
#     print("Accuracy de cada iteración de la validación cruzada: ", accuracy_scores)
#     print("Media de los accuracy: ", accuracy_scores.mean())

## Dataframe Completo sin Edición

### Random Forest

# Uso de la función
# train_and_evaluate_rf(df, 200, 15)

**Validación cruzada**

# evaluate_model_cv_rf(df, 200, 15)

### Extra Trees

# Uso de la función
# train_and_evaluate_et(df, 150, 25,5)

**Validación Cruzada**

# evaluate_model_cv_et(df, 150, 25, 5)

### Análisis

- **Rendimiento en el entrenamiento:** Extra Trees tiene un rendimiento ligeramente mejor en el conjunto de entrenamiento con una precisión del 90% en comparación con el 89% de Random Forest.

- **Rendimiento en la prueba:** Random Forest tiene un rendimiento ligeramente mejor en el conjunto de pruebas con una precisión del 79% en comparación con el 78% de Extra Trees.

- **Validación Cruzada:** Ambos modelos tienen rendimientos similares en la validación cruzada, lo que significa que su rendimiento es bastante consistente a través de diferentes divisiones de los datos.

En este caso se valora más el rendimiento en datos no vistos (conjunto de pruebas). Por lo tanto, Random Forest podría ser una mejor opción.

## Dataframe Completo con Edición

### Random Forest

# Uso de la función
# train_and_evaluate_rf(df_edit, 150, 13)

**Validación cruzada**

# evaluate_model_cv_rf(df_edit, 150, 13)

### Extra Trees

# Uso de la función
# train_and_evaluate_et(df_edit, 150, 22, 5)

**Validación Cruzada**

# evaluate_model_cv_et(df_edit, 150, 22, 5)

### Análisis

Mirando los resultados de ambos modelos, Random Forest y Extra Trees, podemos realizar un análisis en varios aspectos:

- **Precisión general (Accuracy):** Ambos modelos muestran un nivel de precisión similar en el conjunto de entrenamiento con alrededor del 83%. Sin embargo, en el conjunto de prueba, Random Forest tiene un rendimiento ligeramente mejor con una precisión del 72%, mientras que Extra Trees muestra una precisión del 70%.

- **Precisión y Recall por clase:** En el conjunto de entrenamiento, Extra Trees parece tener una mejor precisión para las clases 'cincuenta', 'cuarenta', 'cuarenta y cinco', 'quince', 'treinta' y 'treinta y cinco', pero Random Forest tiene un mejor rendimiento en 'veinte' y 'veinticinco'. Para el conjunto de prueba, vemos una situación similar, aunque las diferencias son menos pronunciadas. Notablemente, Extra Trees parece tener una mayor capacidad para predecir correctamente la clase 'cincuenta' con una precisión y recall altos.

- **F1-Score:** En el conjunto de entrenamiento, ambos modelos muestran un F1-Score promedio similar. Sin embargo, en el conjunto de prueba, Random Forest tiene un rendimiento ligeramente mejor con un F1-Score promedio de 0.71 frente al 0.70 de Extra Trees.

- **Validación Cruzada:** Los resultados de la validación cruzada (Cross-Validation) para Random Forest son consistentemente más altos que para Extra Trees, tanto en términos de precisión (score), F1-Score como recall, sugiriendo que este modelo es más robusto y menos propenso a sobreajuste.

Considerando estos factores, parecería que Random Forest es el modelo superior. Sin embargo, vale la pena señalar que las diferencias entre los dos modelos no son abrumadoras. 

## Dataframe 2017-2022 sin Edición

### Random Forest

# Uso de la función
# train_and_evaluate_rf(df_2017, 200, 17)

**Validación Cruzada**

# evaluate_model_cv_rf(df_2017, 200, 17)

### Extra Trees

# Uso de la función
# train_and_evaluate_et(df_2017, 150, 25, 3)

**Validación Cruzada**

# evaluate_model_cv_et(df_2017, 150, 25, 3)

### Análisis

Ambos modelos, Random Forest y Extra Trees, muestran un rendimiento sólido. Sin embargo, en general, Random Forest parece superar ligeramente a Extra Trees en términos de métricas de rendimiento. Específicamente:

La precisión ponderada en los conjuntos de entrenamiento y prueba es ligeramente mayor en Random Forest.
Las medias de validación cruzada para scores, F1-scores, y recalls son también un poco mejores en Random Forest.
Por lo tanto, mi criterio me inclina a seleccionar el Random Forest como el mejor modelo en este caso, debido a su rendimiento generalmente más fuerte en las métricas clave de clasificación. 

## Dataframe 2017-2022 con Edición

### Random Forest

# Uso de la función
# train_and_evaluate_rf(df_edit_2017, 200, 15)

**Validación Cruzada**

# evaluate_model_cv_rf(df_edit_2017, 200, 17)

### Extra Trees

# # Uso de la función
# train_and_evaluate_et(df_edit_2017, 150, 25, 3)

**Validación Cruzada**

# evaluate_model_cv_et(df_edit_2017, 150, 25, 3)

### Análisis

Los resultados de ambos modelos son bastante similares. Sin embargo, el modelo Random Forest muestra una mejor exactitud en ambos conjuntos de entrenamiento y prueba. Además, el modelo Random Forest también tiene una media de los scores de validación cruzada levemente superior.

Además de la exactitud y los scores de validación cruzada, el F1-score es otra métrica importante que debemos considerar. Es una métrica que combina precisión y recall, y puede ser una mejor métrica cuando las clases están desbalanceadas. El F1-score medio del modelo Random Forest también es superior al del modelo Extra Trees.

Por lo tanto, basándonos en estos resultados, podemos concluir que el modelo Random Forest del conjunto del 2017-2022 con Edición parece ser la mejor opción.

# y = df_edit_2017['dosis_coagulante']
# X = df_edit_2017.drop('dosis_coagulante', axis=1)

# X.columns

# def plot_learning_curve(estimator, title, X, y, ylim=None, cv=None, n_jobs=None, train_sizes=np.linspace(.1, 1.0, 5)):

#     plt.figure()
#     plt.title(title)
#     if ylim is not None:
#         plt.ylim(*ylim)
#     plt.xlabel("Training examples")
#     plt.ylabel("Score")
    
#     train_sizes, train_scores, test_scores = learning_curve(estimator, X, y, cv=cv, n_jobs=n_jobs, train_sizes=train_sizes)

#     train_scores_mean = np.mean(train_scores, axis=1)
#     train_scores_std = np.std(train_scores, axis=1)
#     test_scores_mean = np.mean(test_scores, axis=1)
#     test_scores_std = np.std(test_scores, axis=1)
    
#     plt.grid()

#     plt.fill_between(train_sizes, train_scores_mean - train_scores_std, train_scores_mean + train_scores_std, alpha=0.1, color="r")
#     plt.fill_between(train_sizes, test_scores_mean - test_scores_std, test_scores_mean + test_scores_std, alpha=0.1, color="g")
#     plt.plot(train_sizes, train_scores_mean, 'o-', color="r", label="Training score")
#     plt.plot(train_sizes, test_scores_mean, 'o-', color="g", label="Cross-validation score")

#     plt.legend(loc="best")
#     return plt

# cv = ShuffleSplit(n_splits=30, test_size=0.2, random_state=0)

# estimator = RandomForestClassifier(n_estimators=200, max_depth=17, random_state=42)

# plot_learning_curve(estimator, "Curvas de aprendizaje (Random Forest)", X, y, ylim=(0.6, 1.01), cv=cv, n_jobs=-1)
# plt.show()

Con este modelo se hará una pequeña evaluación para ver que tanta proximidad a los valores reales se tiene.

# def plot_learning_curve_2(estimator, title, X, y, ylim=None, cv=None, n_jobs=None, train_sizes=np.linspace(.1, 1.0, 5)):

#     plt.figure()
#     plt.title(title)
#     if ylim is not None:
#         plt.ylim(*ylim)
#     plt.xlabel("Training examples")
#     plt.ylabel("Score")
    
#     train_sizes, train_scores, test_scores = learning_curve(estimator, X, y, cv=cv, n_jobs=n_jobs, train_sizes=train_sizes)

#     train_scores_mean = np.mean(train_scores, axis=1)
#     train_scores_std = np.std(train_scores, axis=1)
#     test_scores_mean = np.mean(test_scores, axis=1)
#     test_scores_std = np.std(test_scores, axis=1)
    
#     plt.grid()

#     plt.fill_between(train_sizes, train_scores_mean - train_scores_std, train_scores_mean + train_scores_std, alpha=0.1, color="r")
#     plt.fill_between(train_sizes, test_scores_mean - test_scores_std, test_scores_mean + test_scores_std, alpha=0.1, color="g")
#     plt.plot(train_sizes, train_scores_mean, 'o-', color="r", label="Training score")
#     plt.plot(train_sizes, test_scores_mean, 'o-', color="g", label="Cross-validation score")

#     plt.legend(loc="best")
#     return plt

# cv = ShuffleSplit(n_splits=30, test_size=0.2, random_state=0)

# estimator = ExtraTreesClassifier(n_estimators=150, max_depth=25, min_samples_split=5, random_state=42)

# plot_learning_curve_2(estimator, "Learning Curves (Extra Trees)", X, y, ylim=(0.6, 1.01), cv=cv, n_jobs=-1)
# plt.show()

# y = df_2017['dosis_coagulante']
# X = df_2017.drop('dosis_coagulante', axis=1)

# # Definir los valores de los hiperparámetros que quieres probar
# param_grid = {'n_estimators': np.arange(100, 250, 50),
#               'max_depth': np.arange(10, 30, 5),
#               'min_samples_split': np.arange(2, 10, 2),
#              }

# # Realiza la búsqueda en cuadrícula con validación cruzada
# grid_search = GridSearchCV(RandomForestClassifier(), param_grid, cv=5)
# grid_search.fit(X, y)

# # Imprime la mejor combinación de hiperparámetros
# print("Mejores hiperparámetros: ", grid_search.best_params_)

# Define el número de componentes de PCA
# n_components = 6

# # Crea un nuevo clasificador con los mejores hiperparámetros
# clf = make_pipeline(PCA(n_components=n_components), RandomForestClassifier(n_estimators=200, max_depth=15, random_state=42))

# # Entrena el clasificador en el conjunto de entrenamiento
# clf.fit(X_train, y_train)

# # Prueba el clasificador en el conjunto de prueba
# y_pred = clf.predict(X_test)

# # Predicciones en el conjunto de entrenamiento
# train_preds = clf.predict(X_train)
            
# print("Informe de clasificación para el conjunto de entrenamiento:")
# print(classification_report(y_train, train_preds))

# # Imprime un informe de clasificación
# print("Informe de clasificación para el conjunto de prueba:")
# print(classification_report(y_test, y_pred))

# sio.dump(clf, "clf.skops")

# sio.load("./clf.skops", trusted=True)

# def plot_learning_curve(estimator, title, X, y, ylim=None, cv=None, n_jobs=None, train_sizes=np.linspace(.1, 1.0, 5)):

#     plt.figure()
#     plt.title(title)
#     if ylim is not None:
#         plt.ylim(*ylim)
#     plt.xlabel("Training examples")
#     plt.ylabel("Score")
    
#     train_sizes, train_scores, test_scores = learning_curve(estimator, X, y, cv=cv, n_jobs=n_jobs, train_sizes=train_sizes)

#     train_scores_mean = np.mean(train_scores, axis=1)
#     train_scores_std = np.std(train_scores, axis=1)
#     test_scores_mean = np.mean(test_scores, axis=1)
#     test_scores_std = np.std(test_scores, axis=1)
    
#     plt.grid()

#     plt.fill_between(train_sizes, train_scores_mean - train_scores_std, train_scores_mean + train_scores_std, alpha=0.1, color="r")
#     plt.fill_between(train_sizes, test_scores_mean - test_scores_std, test_scores_mean + test_scores_std, alpha=0.1, color="g")
#     plt.plot(train_sizes, train_scores_mean, 'o-', color="r", label="Training score")
#     plt.plot(train_sizes, test_scores_mean, 'o-', color="g", label="Cross-validation score")

#     plt.legend(loc="best")
#     return plt

# cv = ShuffleSplit(n_splits=30, test_size=0.2, random_state=0)

# estimator = RandomForestClassifier(n_estimators=200, max_depth=30, random_state=42)

# plot_learning_curve(estimator, "Learning Curves (Random Forest)", X, y, ylim=(0.7, 1.01), cv=cv, n_jobs=-1)
# plt.show()

def model(df, n_estimators, max_depth , target_col='dosis_coagulante'):
    y = df[target_col]
    X = df.drop(target_col, axis=1)
    
    # Dividir el conjunto de datos en entrenamiento y prueba
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

    # Crear un pipeline para estandarizar los datos y luego aplicar el clasificador
    clf = make_pipeline(RandomForestClassifier(n_estimators=n_estimators, max_depth=max_depth, random_state=42))
    clf.fit(X_train, y_train)
    
    return clf

model = model(df_edit_2017, 200, 17)

def predict_and_display(model, input_df):
    # Asegurarte de que las columnas del dataframe de entrada coinciden con las del modelo
    expected_columns = ['agua_cruda_p_h', 'agua_cruda_color', 'agua_cruda_alcalinidad',
                        'agua_cruda_conductividad', 'vel_viento', 'precipitacion',
                        'temp_humeda', 'clasificador_mensual']
    
    if not all([col in input_df.columns for col in expected_columns]):
        raise ValueError(f"El DataFrame proporcionado debe contener las siguientes columnas: {expected_columns}")
    
    # Usar el modelo para predecir
    predictions = model.predict(input_df)
    
    # Agregar las predicciones al dataframe original para mostrar en una tabla
    input_df['Predicción'] = predictions
    
    # Imprimir el dataframe resultante
    print(input_df)
    return input_df

data = {
    'agua_cruda_p_h': [7.1, 7.2, 7.3, 7.1, 7.3, 7.3, 7.4, 7.4, 7.3],
    'agua_cruda_color': [32, 23, 23, 131, 20, 88, 17, 36, 33],
    'agua_cruda_alcalinidad': [25, 24, 25, 31, 22, 22, 26, 28, 21],
    'agua_cruda_conductividad': [49, 51.9, 48.3, 47.5, 51.4, 48.8, 55.8, 60.9, 47.3],
    'vel_viento': [0.65, 0.908333333, 1, 0.633333333, 2.125, 1.625, 1.083333333, 1.691666666, 1.44166666],
    'precipitacion': [0, 0, 0, 0, 0, 0, 0, 0, 0],
    'temp_humeda': [14, 17, 15, 15, 18, 16, 15, 17, 14],
    'clasificador_mensual': [1, 1, 1, 1, 1, 1, 1, 1, 1]
}

input_df = pd.DataFrame(data)

output_df = predict_and_display(model, input_df)

import pickle

# Guardar el modelo a un archivo
with open('rf_model.pkl', 'wb') as model_file:
    pickle.dump(model, model_file)
